requirements.txt
aiodns==2.0.0
asgiref==3.7.2
cffi==1.16.0
colorama==0.4.6
Django==4.2.3
django-cors-headers==4.2.0
django-countries==7.5.1
django-ipware==6.0.0
django-rest-passwordreset==1.3.0
django-simple-history==3.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
dnspython==2.4.2
drf-recaptcha==3.0.0
drf-yasg==1.21.6
idna==2.10
inflection==0.5.1
iniconfig==2.0.0
mysqlclient==2.2.1
packaging==23.1
Pillow==10.0.0
pluggy==1.3.0
pycares==4.4.0
pycparser==2.21
pydnsbl==1.1.6
PyJWT==2.8.0
pytest==7.4.3
python-dotenv==1.0.0
python-ipware==2.0.0
python-magic==0.4.27
python-magic-bin==0.4.14
pytz==2023.3
PyYAML==6.0.1
setuptools==69.0.2
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
